feat: extracted abstract layer control functionality
